# openapi.yaml
openapi: 3.1.0
info:
  title: Codestra Evolution Provider Adapter
  version: 0.1.0
  description: Internal provider adapter behind Middleware V3. Not a public command API.
paths:
  /healthz:
    get:
      responses:
        "200": { description: Healthy }
  /readyz:
    get:
      responses:
        "200": { description: Internal service authentication and required provider/event configuration are ready }
        "503": { description: Fail-closed readiness problem }
  /internal/v1/whatsapp/transport/health:
    get:
      responses:
        "200": { description: Provider configuration status }
  /internal/v1/whatsapp/transport/messages/{provider_message_id}:
    get:
      description: Internal provider message readback for Middleware V3.
      parameters:
        - in: path
          name: provider_message_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Normalized provider state }
        "401": { description: Service authentication failed }
        "404": { description: Provider message not found }
  /internal/v1/whatsapp/transport/messages:
    post:
      description: Internal Middleware-to-provider translation. Header/body identifiers must match exactly.
      parameters:
        - { name: X-Command-ID, in: header, required: true, schema: { type: string } }
        - { name: X-Tenant-ID, in: header, required: true, schema: { type: string } }
        - { name: X-Correlation-ID, in: header, required: true, schema: { type: string } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      responses:
        "202": { description: Provider accepted }
        "400": { description: Missing required identifiers or invalid JSON }
        "401": { description: Service authentication failed }
        "409": { description: Header/body identifier mismatch }
        "415": { description: JSON content type required }
        "423": { description: Provider effects disabled }
  /internal/v1/whatsapp/webhooks/evolution:
    post:
      responses:
        "202": { description: Provider event normalized }
  /internal/v1/whatsapp/webhooks/meta:
    get:
      responses:
        "200": { description: Meta webhook verification challenge }
        "403": { description: Verification failed }
    post:
      responses:
        "202": { description: Meta event normalized }